Factor out prediction module and add complrex prediction with interpolation
